# api/openapi.yaml
openapi: 3.0.3
info:
  title: Minato Control Plane API
  description: |
    REST API for managing game servers, fleets, profiles, and snapshots.
    All endpoints under `/api/v1` require authentication except where noted.

    This document is the **source of truth** for the API contract
    (see `docs/architecture/api-conventions.md`). Server routes are generated
    from it; edit the spec first, then regenerate.
  version: 1.0.0
  contact:
    name: 7K Group
    url: https://7kgroup.com
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
  - basicAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/v1/gameservers:
    get:
      operationId: listGameServers
      tags: [GameServers]
      summary: List game servers
      description: List all GameServers across all namespaces
      responses:
        "200":
          description: List of game servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameServer"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/gameservers/{namespace}:
    post:
      operationId: createGameServer
      tags: [GameServers]
      summary: Create game server
      description: Create a new GameServer in the specified namespace (admin only)
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameServer"
      responses:
        "201":
          description: Game server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/gameservers/{namespace}/{name}:
    get:
      operationId: getGameServer
      tags: [GameServers]
      summary: Get game server
      description: Get a specific GameServer by namespace and name
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Game server details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateGameServerLifecycle
      tags: [GameServers]
      summary: Patch game server lifecycle
      description: |
        JSON merge-patch for lifecycle control (operator+ only). Only
        `spec.lifecycle.autoStart` and `spec.lifecycle.idleTimeoutSeconds`
        may be changed; any other field is rejected with 400.

        Setting `autoStart: false` on a running server gracefully stops it
        (agent PrepareShutdown, StatefulSet scaled to 0, status.state
        becomes `Stopped`). Setting `autoStart: true` on a stopped server
        starts it again. This is the data-plane stop used by minato-cloud
        to suspend servers (e.g. for dunning).
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameServerLifecyclePatch"
      responses:
        "200":
          description: Updated game server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGameServer
      tags: [GameServers]
      summary: Delete game server
      description: Delete a GameServer (admin only)
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "204":
          description: Game server deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/gameservers/{namespace}/{name}/console:
    get:
      operationId: getConsole
      tags: [Console]
      summary: Console WebSocket stream
      description: |
        Upgrades to a WebSocket connection streaming live console output and
        accepting console commands. The message protocol is defined in
        `docs/agent-developers/console-protocol.md` (typed JSON frames:
        client sends `rcon`/`ping`, server sends `log`/`rcon-response`/`status`/`error`).
        Requires operator role or higher.
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "101":
          description: Switching Protocols — WebSocket stream established
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/gameservers/{namespace}/{name}/actions:
    get:
      operationId: listActions
      tags: [Actions]
      summary: List available actions
      description: List actions available for a game server based on its profile
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: List of available actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Action"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/gameservers/{namespace}/{name}/actions/{action}:
    post:
      operationId: executeAction
      tags: [Actions]
      summary: Execute action
      description: Execute an action on a game server (operator+ only)
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
        - name: action
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
              example:
                message: "Hello players!"
      responses:
        "201":
          description: Action execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionExecutionRef"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/gameservers/{namespace}/{name}/executions/{executionId}:
    get:
      operationId: getActionExecution
      tags: [Actions]
      summary: Get action execution status
      description: Get the status of an action execution
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
        - name: executionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Action execution details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionExecution"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/gameservers/{namespace}/{name}/snapshots:
    get:
      operationId: listSnapshots
      tags: [Snapshots]
      summary: List snapshots
      description: List snapshots for a game server
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: List of snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSnapshot"
    post:
      operationId: createSnapshot
      tags: [Snapshots]
      summary: Create snapshot
      description: Create a snapshot for a game server (operator+ only)
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "201":
          description: Snapshot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/gameserverfleets:
    get:
      operationId: listGameServerFleets
      tags: [Fleets]
      summary: List fleets
      description: List all GameServerFleets across all namespaces
      responses:
        "200":
          description: List of fleets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameServerFleet"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/gameserverfleets/{namespace}:
    post:
      operationId: createGameServerFleet
      tags: [Fleets]
      summary: Create fleet
      description: |
        Create a new GameServerFleet in the specified namespace (admin only).
        The target namespace is auto-created (and hardened) if it does not
        exist. The referenced GameProfile must exist and replicas must be >= 0.
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameServerFleet"
      responses:
        "201":
          description: Fleet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServerFleet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/gameserverfleets/{namespace}/{name}:
    get:
      operationId: getGameServerFleet
      tags: [Fleets]
      summary: Get fleet
      description: Get a specific GameServerFleet
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Fleet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServerFleet"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateGameServerFleet
      tags: [Fleets]
      summary: Scale fleet
      description: |
        JSON merge-patch for fleet scaling (operator+ only). Only
        `spec.replicas` may be changed; any other field is rejected with 400.
        replicas must be >= 0.
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameServerFleetPatch"
      responses:
        "200":
          description: Updated fleet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServerFleet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGameServerFleet
      tags: [Fleets]
      summary: Delete fleet
      description: Delete a GameServerFleet and its GameServers (admin only)
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "204":
          description: Fleet deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/profiles:
    get:
      operationId: listProfiles
      tags: [Profiles]
      summary: List profiles
      description: List all GameProfiles (cluster-scoped)
      responses:
        "200":
          description: List of game profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        OIDC JWT token for humans, or a minato API key (`minato_...`) for
        machine callers. `X-API-Key` is accepted as a legacy alias for API keys.
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication (local development only)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "Legacy alias for API key authentication; prefer `Authorization: Bearer`"
  schemas:
...
    GameServer:
      type: object
      properties:
        apiVersion:
          type: string
          example: operator.minato.io/v1
        kind:
          type: string
          example: GameServer
        metadata:
          $ref: "#/components/schemas/ObjectMeta"
        spec:
          $ref: "#/components/schemas/GameServerSpec"
        status:
          $ref: "#/components/schemas/GameServerStatus"
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...